+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # Changelog
 ## Version 0
 ### Version 0.1.0
-* Intial version
+* Intial version. Includes #conditions, #moon_phase, and #alerts.